refactor toolbar into a seperate directory
