+++ src/components/RichTextEditor/RichTextEditor.tsx
@@ -5,10 +5,10 @@ import { Plate, TNode, usePlateEditorState } from '@udecode/plate';
 import { debounce } from '@utilities/debounce';
 import React, { FC, useCallback, useEffect, useRef, useState } from 'react';
 import { EditableProps } from 'slate-react/dist/components/editable';
-import { Toolbar } from './Toolbar';
+import { Toolbar } from './components/Toolbar/Toolbar';
+import { TextStyleType } from './types';
 import { EditorActions } from './utils/actions';
 import { getEditorConfig } from './utils/getEditorConfig';
-import { TextStyleType } from './utils/getTextStyles';
 import { EMPTY_RICH_TEXT_VALUE, parseRawValue } from './utils/parseRawValue';
 
 export type RichTextEditorProps = {
