+++ src/components/RichTextEditor/components/Toolbar/ButtonGroup.tsx
@@ -32,12 +32,12 @@ import {
     usePlateEditorRef,
 } from '@udecode/plate';
 import React, { FC, ReactElement } from 'react';
-import { ELEMENT_CHECK_ITEM } from './plugins/checkboxListPlugin/createCheckboxListPlugin';
-import { CheckboxListToolbarButton } from './plugins/checkboxListPlugin/ui/CheckboxListToolbarButton';
-import { EditLinkChooserButton } from './plugins/linkChooserPlugin/ui/EditLinkChooserButton';
-import { TextStyleDropdown } from './TextStyleDropdown/TextStyleDropdown';
-import { ButtonGroupProps, IconStylingWrapperProps, TextStyleType } from './types';
-import { EditorActions } from './utils/actions';
+import { ELEMENT_CHECK_ITEM } from '../../plugins/checkboxListPlugin/createCheckboxListPlugin';
+import { CheckboxListToolbarButton } from '../../plugins/checkboxListPlugin/ui/CheckboxListToolbarButton';
+import { EditLinkChooserButton } from '../../plugins/linkChooserPlugin/ui/EditLinkChooserButton';
+import { TextStyleDropdown } from '../../TextStyleDropdown/TextStyleDropdown';
+import { ButtonGroupProps, IconStylingWrapperProps, TextStyleType } from '../../types';
+import { EditorActions } from '../../utils/actions';
 const classNames = {
     root: 'tw-text-text-weak tw-ml-0.5 hover:tw-bg-box-selected hover:!tw-text-box-selected-inverse hover:tw-rounded',
     active: 'tw-bg-box-selected tw-rounded !tw-text-box-selected-inverse',
